Double speed of area mining drills and industrial furnaces.
After careful consideration of the balance, make it so that the higher
tiers of machine are about twice as good speed-over-space wise as the
lower tiers, to encourage using them both with and without modules.

Balance for this change may break with higher tiers of modules, but for
vanilla this feels about right for what I'm looking for.